# docs/api/gateway-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PQC Migration Gateway API",
    "version": "0.1.0",
    "description": "Gateway proxy API for discovery, risk scoring, proof generation, and QASM examples."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/discovery": {
      "post": {
        "summary": "Run discovery scan",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string"
                  },
                  "port": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Discovery response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "target",
                    "findings"
                  ],
                  "properties": {
                    "target": {
                      "$ref": "#/components/schemas/DiscoveryTarget"
                    },
                    "findings": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Finding" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk": {
      "post": {
        "summary": "Compute HNDL risk score",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "total_assets": {
                    "type": "number",
                    "minimum": 0
                  },
                  "quantum_vulnerable_assets": {
                    "type": "number",
                    "minimum": 0
                  },
                  "policy": {
                    "type": "string",
                    "enum": ["strict", "balanced", "lenient"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Risk score response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "policy",
                    "exposure_ratio",
                    "score"
                  ],
                  "properties": {
                    "policy": {
                      "type": "string"
                    },
                    "exposure_ratio": {
                      "type": "number"
                    },
                    "score": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/proof": {
      "post": {
        "summary": "Generate proof payload",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "statement": {
                    "type": "string"
                  },
                  "credit_score": {
                    "type": "integer"
                  },
                  "debt_to_income_bps": {
                    "type": "integer"
                  },
                  "late_payments": {
                    "type": "integer"
                  },
                  "existing_loans": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proof response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statement",
                    "score_value",
                    "score_band",
                    "proof_hash"
                  ],
                  "properties": {
                    "statement": {
                      "type": "string"
                    },
                    "score_value": {
                      "type": "integer"
                    },
                    "score_band": {
                      "type": "string"
                    },
                    "proof_hash": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/qasm": {
      "post": {
        "summary": "List or fetch QASM examples",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "QASM examples response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/QasmExamplesResponse" },
                    { "$ref": "#/components/schemas/QasmSourceResponse" }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/qasm/run": {
      "post": {
        "summary": "Validate and run a named QASM example manifest",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QasmRunRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validated QASM manifest",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/QasmRunResponse" }
              }
            }
          },
          "400": { "description": "Invalid filename, shots, or manifest" }
        }
      }
    },
    "/api/v1/assets": {
      "get": {
        "summary": "List deduplicated discovered assets",
        "responses": {
          "200": {
            "description": "Asset list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "count",
                    "assets"
                  ],
                  "properties": {
                    "count": {
                      "type": "integer"
                    },
                    "assets": { "type": "array", "items": { "$ref": "#/components/schemas/Asset" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create or upsert a manually supplied asset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssetCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing asset upserted",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AssetCreateResponse" } } }
          },
          "201": {
            "description": "New asset created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AssetCreateResponse" } } }
          },
          "400": { "description": "Invalid asset fields" }
        }
      }
    },
    "/api/v1/risk/backlog": {
      "post": {
        "summary": "Export ranked migration backlog",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "policy": {
                    "type": "string",
                    "enum": ["strict", "balanced", "lenient"]
                  },
                  "asset_rows": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/BacklogRow" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranked backlog",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "policy",
                    "backlog"
                  ],
                  "properties": {
                    "policy": {
                      "type": "string"
                    },
                    "backlog": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/BacklogResult" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/governance/exceptions": {
      "get": {
        "summary": "List governance exceptions",
        "responses": {
          "200": {
            "description": "Exception list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "exceptions"
                  ],
                  "properties": {
                    "exceptions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GovernanceException"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create governance exception",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GovernanceExceptionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created exception record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernanceException"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DiscoveryTarget": {
        "type": "object",
        "required": ["address", "port"],
        "properties": {
          "address": { "type": "string" },
          "port": { "type": "integer", "minimum": 1, "maximum": 65535 }
        }
      },
      "Finding": {
        "type": "object",
        "required": ["target", "protocol", "severity", "summary"],
        "properties": {
          "target": { "type": "string" },
          "protocol": { "type": "string" },
          "severity": { "type": "string", "enum": ["critical", "high", "medium", "low", "info", "unknown"] },
          "summary": { "type": "string" }
        }
      },
...
      "QasmExamplesResponse": {
        "type": "object",
        "required": ["examples"],
        "properties": { "examples": { "type": "array", "items": { "$ref": "#/components/schemas/QasmExample" } } }
      },
      "QasmSourceResponse": {
        "type": "object",
        "required": ["name", "source"],
        "properties": { "name": { "type": "string" }, "source": { "type": "string" } }
      },
...
      "AssetCreateRequest": {
        "type": "object",
        "required": ["target", "protocol", "severity", "summary"],
        "properties": {
          "target": { "type": "string", "maxLength": 512 },
          "protocol": { "type": "string", "maxLength": 64 },
          "severity": { "type": "string", "enum": ["critical", "high", "medium", "low", "info"] },
          "summary": { "type": "string", "maxLength": 2048 }
        }
      },
...
      "QasmRunRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "pattern": "^[A-Za-z0-9_.-]+\\.qasm$" },
          "workflow_name": { "type": "string" },
          "backend": { "type": "string", "default": "simulator" },
          "shots": { "type": "integer", "minimum": 1, "maximum": 1000000, "default": 1024 }
        }
      },
      "QasmRunResponse": {
        "type": "object",
        "required": ["workflow_name", "qasm_path", "backend", "shots", "line_count", "manifest_hash", "status"],
        "properties": {
          "workflow_name": { "type": "string" },
          "qasm_path": { "type": "string" },
          "backend": { "type": "string" },
          "shots": { "type": "integer" },
          "line_count": { "type": "integer" },
          "manifest_hash": { "type": "string" },
          "status": { "type": "string", "enum": ["validated"] }
        }
      },
...
    }
  }
}